Make the nested CloudFront component's name a variable.
This is to avoid conflict in the YAML when specifying the config for nested components. If we needed to use this component multiple times with a custom config for the CloudFront component, the 'components->cloudfront' key would be overwritten on merge. This needs to be a unique value to avoid this conflict.